+++ silkopter/brain/src/processor/Gravity_Filter.cpp
@@ -74,7 +74,7 @@ void Gravity_Filter::process()
                           stream::IFrame::Sample const& f_sample,
                           stream::IAcceleration::Sample const& a_sample)
     {
-        auto p2l = math::inverse<float, math::safe>(f_sample.value);
+        auto p2l = math::inverse<double, math::safe>(f_sample.value);
         auto gravity_local = math::rotate(p2l, physics::constants::world_gravity);
         m_output_stream->push_sample(a_sample.value - gravity_local, f_sample.is_healthy & a_sample.is_healthy);
     });
