+++ silkopter/brain/src/sink/PCA9685.h
@@ -73,7 +73,7 @@ private:
 
     std::vector<PWM_Channel> m_pwm_channels;
 
-    void set_pwm_value(size_t idx, boost::optional<float> value);
+    void set_pwm_value(size_t idx, boost::optional<double> value);
 };
 
 }
